BE_CHAT/requirements.txt
Django==5.2.5
djangorestframework==3.15.2
channels==4.1.0
channels-redis==4.2.0
mysqlclient==2.2.4
redis==5.0.8